requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
easyocr==1.7.2
filelock==3.13.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.13.0
Nuitka==2.7.12
numpy==2.2.6
opencv-python-headless==4.12.0.88
ordered-set==4.1.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
pynput==1.8.1
pyparsing==3.2.3
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-bidi==0.6.6
pywin32==311
PyYAML==6.0.2
QtAwesome==1.4.0
QtPy==2.4.3
requests==2.32.5
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.1
setuptools==70.2.0
shapely==2.1.1
shiboken6==6.9.1
six==1.17.0
soupsieve==2.7
sympy==1.13.3
tifffile==2025.6.11
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
zstandard==0.24.0
